Separate apriltag camera into helper class
