+++ src/main/java/frc/robot/subsystems/swervedrive/Vision.java
@@ -1,42 +1,24 @@
 package frc.robot.subsystems.swervedrive;
 
-import static edu.wpi.first.units.Units.Microseconds;
-import static edu.wpi.first.units.Units.Milliseconds;
-import static edu.wpi.first.units.Units.Seconds;
-
 import edu.wpi.first.apriltag.AprilTagFieldLayout;
 import edu.wpi.first.apriltag.AprilTagFields;
-import edu.wpi.first.math.Matrix;
 import edu.wpi.first.math.VecBuilder;
 import edu.wpi.first.math.geometry.Pose2d;
 import edu.wpi.first.math.geometry.Pose3d;
-import edu.wpi.first.math.geometry.Rotation2d;
 import edu.wpi.first.math.geometry.Rotation3d;
 import edu.wpi.first.math.geometry.Transform2d;
-import edu.wpi.first.math.geometry.Transform3d;
 import edu.wpi.first.math.geometry.Translation3d;
-import edu.wpi.first.math.numbers.N1;
-import edu.wpi.first.math.numbers.N3;
 import edu.wpi.first.math.util.Units;
-import edu.wpi.first.networktables.DoubleArrayPublisher;
-import edu.wpi.first.networktables.NetworkTableInstance;
-import edu.wpi.first.networktables.NetworkTablesJNI;
-import edu.wpi.first.wpilibj.Alert;
-import edu.wpi.first.wpilibj.Alert.AlertType;
 import edu.wpi.first.wpilibj.smartdashboard.Field2d;
 import frc.robot.Robot;
+import frc.util.Camera;
 import java.awt.Desktop;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 import java.util.function.Supplier;
 import org.photonvision.EstimatedRobotPose;
-import org.photonvision.PhotonCamera;
-import org.photonvision.PhotonPoseEstimator;
-import org.photonvision.PhotonPoseEstimator.PoseStrategy;
 import org.photonvision.PhotonUtils;
-import org.photonvision.simulation.PhotonCameraSim;
-import org.photonvision.simulation.SimCameraProperties;
 import org.photonvision.simulation.VisionSystemSim;
 import org.photonvision.targeting.PhotonPipelineResult;
 import org.photonvision.targeting.PhotonTrackedTarget;
@@ -49,32 +31,14 @@ import swervelib.telemetry.SwerveDriveTelemetry;
  */
 public class Vision {
 
-  /**
-   * April Tag Field Layout of the year.
-   */
   public static final AprilTagFieldLayout fieldLayout = AprilTagFieldLayout.loadField(
     AprilTagFields.k2025ReefscapeWelded
   );
-  /**
-   * Ambiguity defined as a value between (0,1). Used in {@link Vision#filterPose}.
-   */
-  private final double maximumAmbiguity = 0.25;
-  /**
-   * Photon Vision Simulation
-   */
   public VisionSystemSim visionSim;
-  /**
-   * Count of times that the odom thinks we're more than 10meters away from the april tag.
-   */
-  private double longDistangePoseEstimationCount = 0;
-  /**
-   * Current pose from the pose estimator using wheel odometry.
-   */
   private Supplier<Pose2d> currentPose;
-  /**
-   * Field from {@link swervelib.SwerveDrive#field}
-   */
   private Field2d field2d;
+
+  public Camera[] cameras;
 
   /**
    * Constructor for the Vision class.
@@ -86,11 +50,36 @@ public class Vision {
     this.currentPose = currentPose;
     this.field2d = field;
 
+    Camera leftCamera = new Camera(
+      "flCam2025",
+      new Rotation3d(0, Math.toRadians(0), Math.toRadians(0)),
+      new Translation3d(
+        Units.inchesToMeters(12.25),
+        Units.inchesToMeters(-6.25),
+        Units.inchesToMeters(11.375)
+      ),
+      VecBuilder.fill(.5, .5, 1),
+      VecBuilder.fill(.25, .25, .5)
+    );
+    Camera rightCamera = new Camera(
+      "frCam2025",
+      new Rotation3d(0, Math.toRadians(0), Math.toRadians(0)),
+      new Translation3d(
+        Units.inchesToMeters(12.375),
+        Units.inchesToMeters(-6.375),
+        Units.inchesToMeters(11)
+      ),
+      VecBuilder.fill(.5, .5, 1),
+      VecBuilder.fill(.25, .25, .5)
+    );
+    cameras[0] = leftCamera;
+    cameras[1] = rightCamera;
+
     if (Robot.isSimulation()) {
       visionSim = new VisionSystemSim("Vision");
       visionSim.addAprilTags(fieldLayout);
 
-      for (Cameras c : Cameras.values()) {
+      for (Camera c : cameras) {
         c.addToVisionSim(visionSim);
       }
 
@@ -139,7 +128,7 @@ public class Vision {
        */
       visionSim.update(swerveDrive.getSimulationDriveTrainPose().get());
     }
-    for (Cameras camera : Cameras.values()) {
+    for (Camera camera : cameras) {
       Optional<EstimatedRobotPose> poseEst = getEstimatedGlobalPose(camera);
       if (poseEst.isPresent()) {
         var pose = poseEst.get();
@@ -163,7 +152,7 @@ public class Vision {
    *
    * @return an {@link EstimatedRobotPose} with an estimated pose, timestamp, and targets used to create the estimate
    */
-  public Optional<EstimatedRobotPose> getEstimatedGlobalPose(Cameras camera) {
+  public Optional<EstimatedRobotPose> getEstimatedGlobalPose(Camera camera) {
     Optional<EstimatedRobotPose> poseEst = camera.getEstimatedGlobalPose();
     if (Robot.isSimulation()) {
       Field2d debugField = visionSim.getDebugField();
@@ -179,52 +168,6 @@ public class Vision {
       );
     }
     return poseEst;
-  }
-
-  /**
-   * Filter pose via the ambiguity and find best estimate between all of the camera's throwing out distances more than
-   * 10m for a short amount of time.
-   *
-   * @param pose Estimated robot pose.
-   * @return Could be empty if there isn't a good reading.
-   */
-  @Deprecated(since = "2024", forRemoval = true)
-  private Optional<EstimatedRobotPose> filterPose(
-    Optional<EstimatedRobotPose> pose
-  ) {
-    if (pose.isPresent()) {
-      double bestTargetAmbiguity = 1; // 1 is max ambiguity
-      for (PhotonTrackedTarget target : pose.get().targetsUsed) {
-        double ambiguity = target.getPoseAmbiguity();
-        if (ambiguity != -1 && ambiguity < bestTargetAmbiguity) {
-          bestTargetAmbiguity = ambiguity;
-        }
-      }
-      //ambiguity to high dont use estimate
-      if (bestTargetAmbiguity > maximumAmbiguity) {
-        return Optional.empty();
-      }
-
-      //est pose is very far from recorded robot pose
-      if (
-        PhotonUtils.getDistanceToPose(
-          currentPose.get(),
-          pose.get().estimatedPose.toPose2d()
-        ) >
-        1
-      ) {
-        longDistangePoseEstimationCount++;
-
-        //if it calculates that were 10 meter away for more than 10 times in a row its probably right
-        if (longDistangePoseEstimationCount < 10) {
-          return Optional.empty();
-        }
-      } else {
-        longDistangePoseEstimationCount = 0;
-      }
-      return pose;
-    }
-    return Optional.empty();
   }
 
   /**
@@ -249,7 +192,7 @@ public class Vision {
    * @param camera Camera to check.
    * @return Tracked target.
    */
-  public PhotonTrackedTarget getTargetFromId(int id, Cameras camera) {
+  public PhotonTrackedTarget getTargetFromId(int id, Camera camera) {
     PhotonTrackedTarget target = null;
     for (PhotonPipelineResult result : camera.resultsList) {
       if (result.hasTargets()) {
@@ -297,7 +240,7 @@ public class Vision {
    */
   public void updateVisionField() {
     List<PhotonTrackedTarget> targets = new ArrayList<PhotonTrackedTarget>();
-    for (Cameras c : Cameras.values()) {
+    for (Camera c : cameras) {
       if (!c.resultsList.isEmpty()) {
         PhotonPipelineResult latest = c.resultsList.get(0);
         if (latest.hasTargets()) {
@@ -318,391 +261,5 @@ public class Vision {
     }
 
     field2d.getObject("tracked targets").setPoses(poses);
-  }
-
-  /**
-   * Camera Enum to select each camera
-   */
-  public enum Cameras {
-    /**
-     * Left Camera
-     */
-    LEFT_CAM(
-      "flCam2025",
-      new Rotation3d(0, Math.toRadians(0), Math.toRadians(0)),
-      new Translation3d(
-        Units.inchesToMeters(15.508),
-        Units.inchesToMeters(-6.25),
-        Units.inchesToMeters(9.75)
-      ),
-      VecBuilder.fill(.5, .5, 1),
-      VecBuilder.fill(.25, .25, .5)
-    ),
-    /**
-     * Right Camera
-     */
-    RIGHT_CAM(
-      "frCam2025",
-      new Rotation3d(0, Math.toRadians(0), Math.toRadians(0)),
-      new Translation3d(
-        Units.inchesToMeters(15.508),
-        Units.inchesToMeters(6.25),
-        Units.inchesToMeters(9.75)
-      ),
-      VecBuilder.fill(.5, .5, 1),
-      VecBuilder.fill(.25, .25, .5)
-    );
-
-    //BACK_CAM("bkCam2025",
-    //          new Rotation3d(0, Math.toRadians(180), Math.toRadians(0)),
-    //          new Translation3d(Units.inchesToMeters(-15.508),
-    //                            Units.inchesToMeters(.25),
-    //                            Units.inchesToMeters(9)),
-    //          VecBuilder.fill(4, 4, 8), VecBuilder.fill(0.5, 0.5, 1));
-    /*
-     * Right Camera
-     */
-    //RIGHT_CAM("bkCam2025",
-    //          new Rotation3d(0, Math.toRadians(0), Math.toRadians(0)),
-    //          new Translation3d(Units.inchesToMeters(-2),
-    //                            Units.inchesToMeters(-12.018),
-    //                            Units.inchesToMeters(0)),
-    //          VecBuilder.fill(4, 4, 8), VecBuilder.fill(0.5, 0.5, 1));
-    /**
-     * Back Camera
-     */
-
-    /**
-     * Latency alert to use when high latency is detected.
-     */
-    public final Alert latencyAlert;
-    /**
-     * Camera instance for comms.
-     */
-    public final PhotonCamera camera;
-    /**
-     * Pose estimator for camera.
-     */
-    public final PhotonPoseEstimator poseEstimator;
-    /**
-     * Standard Deviation for single tag readings for pose estimation.
-     */
-    private final Matrix<N3, N1> singleTagStdDevs;
-    /**
-     * Standard deviation for multi-tag readings for pose estimation.
-     */
-    private final Matrix<N3, N1> multiTagStdDevs;
-    /**
-     * Transform of the camera rotation and translation relative to the center of the robot
-     */
-    private final Transform3d robotToCamTransform;
-    /**
-     * Current standard deviations used.
-     */
-    public Matrix<N3, N1> curStdDevs;
-    /**
-     * Estimated robot pose.
-     */
-    public Optional<EstimatedRobotPose> estimatedRobotPose = Optional.empty();
-
-    public Transform3d estimateTagPose;
-
-    /**
-     * Simulated camera instance which only exists during simulations.
-     */
-    public PhotonCameraSim cameraSim;
-    /**
-     * Results list to be updated periodically and cached to avoid unnecessary queries.
-     */
-    public List<PhotonPipelineResult> resultsList = new ArrayList<>();
-    /**
-     * Last read from the camera timestamp to prevent lag due to slow data fetches.
-     */
-    private double lastReadTimestamp = Microseconds
-      .of(NetworkTablesJNI.now())
-      .in(Seconds);
-
-    /**
-     * Construct a Photon Camera class with help. Standard deviations are fake values, experiment and determine
-     * estimation noise on an actual robot.
-     *
-     * @param name                  Name of the PhotonVision camera found in the PV UI.
-     * @param robotToCamRotation    {@link Rotation3d} of the camera.
-     * @param robotToCamTranslation {@link Translation3d} relative to the center of the robot.
-     * @param singleTagStdDevs      Single AprilTag standard deviations of estimated poses from the camera.
-     * @param multiTagStdDevsMatrix Multi AprilTag standard deviations of estimated poses from the camera.
-     *
-     */
-    private final NetworkTableInstance networkTable = NetworkTableInstance
-      .getDefault()
-      .getTable("Vision")
-      .getInstance();
-    private final DoubleArrayPublisher time;
-
-    Cameras(
-      String name,
-      Rotation3d robotToCamRotation,
-      Translation3d robotToCamTranslation,
-      Matrix<N3, N1> singleTagStdDevs,
-      Matrix<N3, N1> multiTagStdDevsMatrix
-    ) {
-      time = networkTable.getDoubleArrayTopic(name).publish();
-      latencyAlert =
-        new Alert(
-          "'" + name + "' Camera is experiencing high latency.",
-          AlertType.kWarning
-        );
-
-      camera = new PhotonCamera(name);
-
-      // https://docs.wpilib.org/en/stable/docs/software/basic-programming/coordinate-system.html
-      robotToCamTransform =
-        new Transform3d(robotToCamTranslation, robotToCamRotation);
-
-      poseEstimator =
-        new PhotonPoseEstimator(
-          Vision.fieldLayout,
-          PoseStrategy.MULTI_TAG_PNP_ON_COPROCESSOR,
-          robotToCamTransform
-        );
-      poseEstimator.setMultiTagFallbackStrategy(PoseStrategy.LOWEST_AMBIGUITY);
-
-      this.singleTagStdDevs = singleTagStdDevs;
-      this.multiTagStdDevs = multiTagStdDevsMatrix;
-
-      if (Robot.isSimulation()) {
-        SimCameraProperties cameraProp = new SimCameraProperties();
-        // A 640 x 480 camera with a 100 degree diagonal FOV.
-        cameraProp.setCalibration(960, 720, Rotation2d.fromDegrees(100));
-        // Approximate detection noise with average and standard deviation error in pixels.
-        cameraProp.setCalibError(0.25, 0.08);
-        // Set the camera image capture framerate (Note: this is limited by robot loop rate).
-        cameraProp.setFPS(30);
-        // The average and standard deviation in milliseconds of image data latency.
-        cameraProp.setAvgLatencyMs(35);
-        cameraProp.setLatencyStdDevMs(5);
-
-        cameraSim = new PhotonCameraSim(camera, cameraProp);
-        cameraSim.enableDrawWireframe(true);
-      }
-    }
-
-    /**
-     * Add camera to {@link VisionSystemSim} for simulated photon vision.
-     *
-     * @param systemSim {@link VisionSystemSim} to use.
-     */
-    public void addToVisionSim(VisionSystemSim systemSim) {
-      if (Robot.isSimulation()) {
-        systemSim.addCamera(cameraSim, robotToCamTransform);
-      }
-    }
-
-    /**
-     * Get the result with the least ambiguity from the best tracked target within the Cache. This may not be the most
-     * recent result!
-     *
-     * @return The result in the cache with the least ambiguous best tracked target. This is not the most recent result!
-     */
-    public Optional<PhotonPipelineResult> getBestResult() {
-      if (resultsList.isEmpty()) {
-        return Optional.empty();
-      }
-
-      PhotonPipelineResult bestResult = resultsList.get(0);
-      double amiguity = bestResult.getBestTarget().getPoseAmbiguity();
-      double currentAmbiguity = 0;
-      for (PhotonPipelineResult result : resultsList) {
-        currentAmbiguity = result.getBestTarget().getPoseAmbiguity();
-        if (currentAmbiguity < amiguity && currentAmbiguity > 0) {
-          bestResult = result;
-          amiguity = currentAmbiguity;
-        }
-      }
-      return Optional.of(bestResult);
-    }
-
-    /**
-     * Get the latest result from the current cache.
-     *
-     * @return Empty optional if nothing is found. Latest result if something is there.
-     */
-    public Optional<PhotonPipelineResult> getLatestResult() {
-      return resultsList.isEmpty()
-        ? Optional.empty()
-        : Optional.of(resultsList.get(0));
-    }
-
-    /**
-     * Get the estimated robot pose. Updates the current robot pose estimation, standard deviations, and flushes the
-     * cache of results.
-     *
-     * @return Estimated pose.
-     */
-    public Optional<EstimatedRobotPose> getEstimatedGlobalPose() {
-      updateUnreadResults();
-      return estimatedRobotPose;
-    }
-
-    /**
-     * Update the latest results, cached with a maximum refresh rate of 1req/15ms. Sorts the list by timestamp.
-     */
-    private void updateUnreadResults() {
-      double mostRecentTimestamp = resultsList.isEmpty()
-        ? 0.0
-        : resultsList.get(0).getTimestampSeconds();
-      double currentTimestamp = Microseconds
-        .of(NetworkTablesJNI.now())
-        .in(Seconds);
-      double debounceTime = Milliseconds.of(15).in(Seconds);
-      time.set(
-        new double[] {
-          mostRecentTimestamp,
-          currentTimestamp,
-          currentTimestamp - mostRecentTimestamp,
-          debounceTime,
-        }
-      );
-
-      //Timer.getFPGATimestamp();
-      for (PhotonPipelineResult result : resultsList) {
-        mostRecentTimestamp =
-          Math.max(mostRecentTimestamp, result.getTimestampSeconds());
-        getClosestReefTag(result);
-      }
-      resultsList =
-        Robot.isReal()
-          ? camera.getAllUnreadResults()
-          : cameraSim.getCamera().getAllUnreadResults();
-      lastReadTimestamp = currentTimestamp;
-      resultsList.sort((PhotonPipelineResult a, PhotonPipelineResult b) -> {
-        return a.getTimestampSeconds() >= b.getTimestampSeconds() ? 1 : -1;
-      });
-
-      if (!resultsList.isEmpty()) {
-        updateEstimatedGlobalPose();
-      }
-    }
-
-    /**
-     * The latest estimated robot pose on the field from vision data. This may be empty. This should only be called once
-     * per loop.
-     *
-     * <p>Also includes updates for the standard deviations, which can (optionally) be retrieved with
-     * {@link Cameras#updateEstimationStdDevs}
-     *
-     * @return An {@link EstimatedRobotPose} with an estimated pose, estimate timestamp, and targets used for
-     * estimation.
-     */
-    private void updateEstimatedGlobalPose() {
-      Optional<EstimatedRobotPose> visionEst = Optional.empty();
-      for (var change : resultsList) {
-        visionEst = poseEstimator.update(change);
-        updateEstimationStdDevs(visionEst, change.getTargets());
-      }
-      estimatedRobotPose = visionEst;
-    }
-
-    /**
-     * Calculates new standard deviations This algorithm is a heuristic that creates dynamic standard deviations based
-     * on number of tags, estimation strategy, and distance from the tags.
-     *
-     * @param estimatedPose The estimated pose to guess standard deviations for.
-     * @param targets       All targets in this camera frame
-     */
-    private void updateEstimationStdDevs(
-      Optional<EstimatedRobotPose> estimatedPose,
-      List<PhotonTrackedTarget> targets
-    ) {
-      if (estimatedPose.isEmpty()) {
-        // No pose input. Default to single-tag std devs
-        curStdDevs = singleTagStdDevs;
-      } else {
-        // Pose present. Start running Heuristic
-        var estStdDevs = singleTagStdDevs;
-        int numTags = 0;
-        double avgDist = 0;
-
-        // Precalculation - see how many tags we found, and calculate an average-distance metric
-        for (var tgt : targets) {
-          var tagPose = poseEstimator
-            .getFieldTags()
-            .getTagPose(tgt.getFiducialId());
-          if (tagPose.isEmpty()) {
-            continue;
-          }
-          numTags++;
-          avgDist +=
-            tagPose
-              .get()
-              .toPose2d()
-              .getTranslation()
-              .getDistance(
-                estimatedPose.get().estimatedPose.toPose2d().getTranslation()
-              );
-        }
-
-        if (numTags == 0) {
-          // No tags visible. Default to single-tag std devs
-          curStdDevs = singleTagStdDevs;
-        } else {
-          // One or more tags visible, run the full heuristic.
-          avgDist /= numTags;
-          // Decrease std devs if multiple targets are visible
-          if (numTags > 1) {
-            estStdDevs = multiTagStdDevs;
-          }
-          // Increase std devs based on (average) distance
-          if (numTags == 1 && avgDist > 4) {
-            estStdDevs =
-              VecBuilder.fill(
-                Double.MAX_VALUE,
-                Double.MAX_VALUE,
-                Double.MAX_VALUE
-              );
-          } else {
-            estStdDevs = estStdDevs.times(1 + (avgDist * avgDist / 30));
-          }
-          curStdDevs = estStdDevs;
-        }
-      }
-    }
-
-    public Transform3d getEstimateTagPose() {
-      return estimateTagPose;
-    }
-
-    private void getClosestReefTag(PhotonPipelineResult result) {
-      Transform3d closestTag = null;
-      int tagID = -1;
-      //int targetTagID = -1;
-      double minDistance = 100.0;
-
-      result.targets.removeIf(tag -> {
-        int tagNum = tag.fiducialId;
-        double maxDistance = 6.0;
-        Transform3d transform = tag.getBestCameraToTarget();
-        return (
-          (transform.getX() > maxDistance || transform.getY() > maxDistance) ||
-          !((tagNum >= 17 && tagNum <= 22) || (tagNum >= 6 && tagNum <= 11))
-        );
-      });
-      for (var tag : result.getTargets()) {
-        tagID = tag.fiducialId;
-        if ((tagID >= 17 && tagID <= 22) || (tagID >= 6 && tagID <= 11)) {
-          Transform3d transform = tag.getBestCameraToTarget();
-          double distance = Math.sqrt(
-            Math.pow(transform.getX(), 2) + Math.pow(transform.getY(), 2)
-          );
-          if (distance < minDistance) {
-            closestTag = transform;
-            minDistance = distance;
-            //targetTagID = tagID;
-          }
-        }
-      }
-      //System.out.println(targetTagID);
-      estimateTagPose = closestTag;
-    }
   }
 }
